Rename DPS side of followers to trackers
