+++ service/tracker/consensus.go
@@ -12,7 +12,7 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-package follower
+package tracker
 
 import (
 	"errors"
